requirements.txt
evalutils==0.2.4
opencxr==1.2.0
numpy==2.1.2
pandas==2.2.3
Pillow==11.0.0
scipy==1.14.1
scikit-image
torch==2.4.1
torchvision==0.19.1
